- Refined confirm diaglogs in pawn shop page

+++ server/contracts/IOcxPriceOracle.sol
@@ -7,7 +7,7 @@ pragma solidity ^0.8.0;
  */
 interface IOcxPriceOracle {
     function setSubmitFee(uint64 _applicationFee, uint64 _valuationFee) external;
-    function getSubmitFee() external view returns (uint256 _applicationFee, uint256 _valuationFee);
+    function getSubmitFee() external view returns (uint64 _applicationFee, uint64 _valuationFee);
     function getBtcUsdPrice() external view returns (int256 _price);
     function getEthUsdPrice() external view returns (int256 _price);
     function getBtcEthPrice() external view returns (int256 _price);
